requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
joblib==1.4.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
altair==5.5.0
